MCO-310: delete the stage-based wizard
Unreachable since the single-page form replaced it: the POST /stage route and
its handler, `draftWizardPage`, `wizardStageContent`, `wizardNavigation`,
`wizardProgress`/`wizardProgressHtml`, and `mergeStageIntoDraft` — which only
existed to preview a half-saved stage. Also drops `draftCategoryFields`, the
combined picker+schema helper the form no longer uses now that it places the
two halves in different sections.

`DraftWizardStage` survives, moved to its own file. The screens are gone but
the grouping still organises parsing and validation — `buildStageJson` and
`ValidateStageStep` both dispatch on it, and a draft row still records one in
`current_stage`. Its `next()`/`previous()` are gone with the navigation: nothing
steps through stages any more.

The two tests for the stage endpoint: one covered stage navigation and went
with it, the other covered specs surviving a JSON round trip — including that
the form's trailing blank row must not become an entry — which still matters,
so it moved to POST /save.

Also corrected a comment calling the category/author/version fragment
endpoints "legacy ... used by old wizard fields". They are what the current
form swaps in.
